Ajustando testes.

+++ src/main/java/br/com/alura/domain/Agencia.java
@@ -11,7 +11,7 @@ public class Agencia {
     }
 
     public Agencia(
-            Integer id,
+            Long id,
             String nome,
             String razaoSocial,
             String cnpj,
@@ -25,7 +25,7 @@ public class Agencia {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
 
     @Column(name = "razao_social")
@@ -41,7 +41,7 @@ public class Agencia {
     @JoinColumn(name = "endereco_id")
     private Endereco endereco;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
